// src/app/shared/datepicker/models/dateModel.ts
export interface INDateD {
    M1: number;
    M2: number;
    M3: number;
    M4: number;
    M5: number;
    M6: number;
    M7: number;
    M8: number;
    M9: number;
    M10: number;
    M11: number;
    M12: number;
    NYear: number;
}

export interface IBSADCal {
    NYear: number;
    StartDate: any;
    EndDate: any;
}

export interface INMothCal {
    Month: number;
    Days: number;
}

export interface IMonthsEngNep {
    Id: number;
    EngName: string;
    NepName: string;
}